refactor tomtom etl/merge
this refactor simplifies and resolves issues in read.dbf() where ID fields were being cast to numeric and being rounded. it solves the latter issue by exporting ogr2ogr bash commands to translate and merge the dbfs, which are executed from R. it simplifies the process by working from abbreviations in the metadata and then scanning for expected files rather than scanning the list of files and deriving the file types from extensions.